+++ Editor/Generatable/GeneratableGetterProperty.cs
@@ -11,10 +11,16 @@ namespace NPTP.UnitySourceGen.Editor.Generatable
         private static Type PropertyType => typeof(T);
         
         private readonly GeneratableField accessedField;
+        private readonly T getterValue;
 
         internal GeneratableGetterProperty(string name, AccessModifier getModifier, bool isStatic, GeneratableField accessedField) : base(name, getModifier, isStatic)
         {
             this.accessedField = accessedField;
+        }
+
+        internal GeneratableGetterProperty(string name, AccessModifier accessModifier, bool isStatic, T getterValue) : base(name, accessModifier, isStatic)
+        {
+            this.getterValue = getterValue;
         }
 
         public override string GenerateStringRepresentation()
@@ -26,9 +32,12 @@ namespace NPTP.UnitySourceGen.Editor.Generatable
             property.Append(SPACE + GetTypeName(PropertyType));
             property.Append(SPACE + Name);
             property.Append(SPACE + GETTER_ARROW);
-            property.Append(SPACE + accessedField.Name);
+            
+            if (accessedField != null) property.Append(SPACE + accessedField.Name);
+            else if (getterValue != null) property.Append(SPACE + GetValueAsString(typeof(T), getterValue));
+            
             property.Append(SEMICOLON);
-                                    
+
             return property.ToString();
         }
     }