+++ Editor/SourceGen.cs
@@ -1,8 +1,10 @@
 using System;
+using System.IO;
 using NPTP.UnitySourceGen.Editor.Enums;
 using NPTP.UnitySourceGen.Editor.Extensions;
 using NPTP.UnitySourceGen.Editor.Generatable;
 using NPTP.UnitySourceGen.Editor.UnitySourceGen.Editor.ScriptWriting;
+using UnityEngine;
 
 namespace NPTP.UnitySourceGen.Editor
 {
@@ -11,6 +13,11 @@ namespace NPTP.UnitySourceGen.Editor
         public static GeneratableClass NewClass(string name, AccessModifier accessModifier) => new GeneratableClass(name, accessModifier, isStatic: false);
         public static GeneratableClass NewStaticClass(string name, AccessModifier accessModifier) => new GeneratableClass(name, accessModifier, isStatic: true);
         public static GeneratableEnum NewEnum(string name, AccessModifier accessModifier) => new GeneratableEnum(name, accessModifier);
+
+        public static bool WriteClassToAssetsScriptFile(string pathInsideAssets, GeneratableClass generatableClass)
+        {
+            return ScriptWriter.TryWrite(Application.dataPath + Path.DirectorySeparatorChar + pathInsideAssets, generatableClass.GenerateStringRepresentation());
+        }
         
         public static bool ReplaceClassInAssetsScriptFile(Type classType, GeneratableClass generatableClass)
         {