+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Unity Source Gen
 ## Changelog
 
+0.1.1
+- SerializedProperty generation
+- Write to file directly from SourceGen
+
 0.1.0
 - Fix a few bugs
 - Can now write directly to files & replace specific classes in script files